Split event_imp into multiple files
